convert children to fixed if necessary
